Refactor subject literal narrowing to separate util file
